S3/Encryptions/helpers/requirements.txt
pycrypto==2.6.1 